refactor(clients): Add GitClient and GitHubClient to public API
- Add GitClient and GitHubClient to clients/__init__.py exports
- Update 4 analysis module imports to use public API path
- Fix test mock paths to patch vibe3.clients.GitClient

Changes:
- src/vibe3/clients/__init__.py: Add GitClient and GitHubClient to __all__
- src/vibe3/analysis/snapshot_service.py: Use public import
- src/vibe3/analysis/serena_service.py: Merge GitClient import with SerenaClient
- src/vibe3/analysis/coverage_service.py: Use public import in lazy import
- src/vibe3/analysis/pre_push_scope.py: Use public import in lazy import
- tests/vibe3/analysis/test_coverage_service.py: Update mock path
- tests/vibe3/analysis/test_pre_push_scope.py: Update mock path

Refs: #1704
