+++ src/vibe3/analysis/serena_service.py
@@ -9,8 +9,7 @@ from typing import TYPE_CHECKING
 from loguru import logger
 
 from vibe3.analysis.serena_file_analyzer import analyze_files
-from vibe3.clients import SerenaClient
-from vibe3.clients.git_client import GitClient
+from vibe3.clients import GitClient, SerenaClient
 from vibe3.exceptions import SerenaError
 from vibe3.models.change_source import ChangeSource
