+++ src/vibe3/analysis/snapshot_service.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from loguru import logger
 
 from vibe3.analysis import dag_service, structure_service
-from vibe3.clients.git_client import GitClient
+from vibe3.clients import GitClient
 from vibe3.exceptions import VibeError
 from vibe3.models import (
     DependencyEdge,
